==> 03_tables.Rmd <==
---
title: "Tables"
date: "Last knitted on `r format(Sys.Date(), '%d %b %Y')`"
author: "Leavitt, Kopf, Weber, Pearson"
output:
  html_document: 
    df_print: paged
    number_sections: yes
    toc: yes
    toc_float: true
    toc_depth: 3
    code_folding: show
    standalone: true
editor_options:
  chunk_output_type: console # switch to 'inline' if preferred
---

# Setup

```{r "setup", message = FALSE}
# packages
library(rlang)
library(tidyverse) # general data wrangling and plotting

# custom funtions
source("scripts/error_propagation.R")
source("scripts/table_functions.R")

# global knitting options for code rendering
knitr::opts_chunk$set(collapse = TRUE, comment = "#>")
```

> NOTE: please knit the `01_calculations.Rmd` and `02_model.Rmd` notebooks first to prepare the data used for generating the tables.


# Load Data

```{r, eval = TRUE}
# load data
compounds <- readxl::read_excel("data/compounds.xlsx")
metadata <- readxl::read_excel("data/metadata.xlsx")
samples <- readr::read_rds("cache/samples_w_epsilons.rds") 
samples_w_epsilons <- readr::read_rds("cache/samples_w_epsilons.rds")
samples_sum <- readr::read_rds("cache/samples_sum.rds")
samples_ab_weighted_sum <- readr::read_rds("cache/samples_ab_weighted_sum.rds")
ring_enrichments <- readr::read_rds("cache/ring_enrichments.rds")
fits_wide <- readr::read_rds("cache/fits_wide.rds")
```

# Table 1: biphytanes data summary

```{r}
# save in data output
table_1 <- samples_sum  %>%
  left_join(compounds, by = "compound") %>%
  left_join(metadata, by = c("organism", "growth")) %>%
  left_join(ring_enrichments, by = c("organism", "growth", "compound")) %>%
  arrange(TD.hours, compound) %>%
  rename(d2H.permil = d2H_weighted_mean.permil, eps.permil = eps_weighted_mean.permil) %>%
  pivot_wider(
    id_cols = c(organism, growth, TD.hours, n_analyses, d2H_water.permil),
    values_from = c(
      starts_with("d2H.permil"), starts_with("d2H_error"),
      starts_with("eps"), starts_with("rel_abundance"), 
      D_eps_per_ring),
    names_from = bp_short,
    names_glue = "{bp_short}: {.value}"
  ) %>%
  left_join(samples_ab_weighted_sum, by = c("organism", "growth")) %>%
  select(TD.hours, n_analyses, d2H_water.permil, 
         starts_with("BP-0"), starts_with("BP-1"), starts_with("BP-2"), starts_with("BP-3"),
         eps_ab_weighted_mean.permil, eps_ab_weighted_mean_error.permil
  )

table_1_formatted <-
  table_1 %>% 
  transmute(
    TD.hours = sprintf("%.1f", TD.hours),
    n_analyses = n_analyses,
    d2H_water.permil = sprintf("%.1f", d2H_water.permil),
    `BP-0: d2H.permil` = sprintf("%.0f \U00B1 %.0f", `BP-0: d2H.permil`,  `BP-0: d2H_error.permil`),
    `BP-0: eps.permil` = sprintf("%.0f \U00B1 %.0f", `BP-0: eps.permil`,  `BP-0: eps_error.permil`),
    `BP-1: d2H.permil` = sprintf("%.0f \U00B1 %.0f", `BP-1: d2H.permil`,  `BP-1: d2H_error.permil`),
    `BP-1: eps.permil` = sprintf("%.0f \U00B1 %.0f", `BP-1: eps.permil`,  `BP-1: eps_error.permil`),
    `BP-1: Deps.permil` = sprintf("%.1f", `BP-1: D_eps_per_ring`),
    `BP-2: d2H.permil` = sprintf("%.0f \U00B1 %.0f", `BP-2: d2H.permil`,  `BP-2: d2H_error.permil`),
    `BP-2: eps.permil` = sprintf("%.0f \U00B1 %.0f", `BP-2: eps.permil`,  `BP-2: eps_error.permil`),
    `BP-2: Deps.permil` = sprintf("%.1f", `BP-2: D_eps_per_ring`),
    `BP-3: d2H.permil` = sprintf("%.0f \U00B1 %.0f", `BP-3: d2H.permil`,  `BP-3: d2H_error.permil`),
    `BP-3: eps.permil` = sprintf("%.0f \U00B1 %.0f", `BP-3: eps.permil`,  `BP-3: eps_error.permil`),
    `BP-3: Deps.permil` = sprintf("%.1f", `BP-3: D_eps_per_ring`),
    `BP-0: rel_abundance` = sprintf("%.0f \U00B1 %.0f", 100 * `BP-0: rel_abundance_mean`, 100 * `BP-0: rel_abundance_error`),
    `BP-1: rel_abundance` = sprintf("%.0f \U00B1 %.0f", 100 * `BP-1: rel_abundance_mean`, 100 * `BP-1: rel_abundance_error`),
    `BP-2: rel_abundance` = sprintf("%.0f \U00B1 %.0f", 100 * `BP-2: rel_abundance_mean`, 100 * `BP-2: rel_abundance_error`),
    `BP-3: rel_abundance` = sprintf("%.0f \U00B1 %.0f", 100 * `BP-3: rel_abundance_mean`, 100 * `BP-3: rel_abundance_error`),
    eps_ab_weighted_mean = sprintf("%.0f \U00B1 %.0f", eps_ab_weighted_mean.permil, eps_ab_weighted_mean_error.permil)
  )

export_to_excel(`numbers` = table_1, `formatted` = table_1_formatted, file = "output/table_1_data_summary.xlsx")
```

## Range and means

```{r}
# range and means
message("Fractionation factors range and means: ")
samples_sum %>%
  summarise(
    min_eps = min(eps_weighted_mean.permil), 
    max_eps = max(eps_weighted_mean.permil), 
    mean_eps = mean(eps_weighted_mean.permil), 
    sd = sd(eps_weighted_mean.permil) %>% round()
  )

# mean ring enrichment
message("Mean ring enrichments: ")
ring_enrichments %>% 
  summarize(
    mean_ring_enrichment = mean(D_eps_per_ring) %>% round(1), 
    sd = sd(D_eps_per_ring) %>% round(1)
  )
```

# Table 3: Ring effects

```{r}
# table 3 (model calculation summary)
table_3 <- readr::read_rds("cache/H_accounting.rds") %>%
  mutate(
    across(n_A_W:n_x, 
           function(x) x/n_BP, 
           .names = "f_{str_remove(.col, 'n_')}"
    )
  ) %>%
  group_by(scenario) %>%
  mutate(
    Df_w_and_x = (f_W + f_x) - (f_W[bp_short == "BP-0"] + f_x[bp_short == "BP-0"]),
    Df_NADPH = f_NADPH - f_NADPH[bp_short == "BP-0"],
    Df_FAD = f_FAD - f_FAD[bp_short == "BP-0"]
  ) %>%
  ungroup() %>%
  filter(n_rings > 0) %>%
  select(scenario, bp_short, starts_with("Df"))

# export table S3
table_3_formatted <-
  table_3 %>%
  transmute(
    scenario = scenario, 
    bp = str_remove(bp_short, "-"),
    across(
      starts_with("Df"), 
      function(x) sprintf("%+.2f%%", 100 * x), 
      .names = "f_{str_remove(.col, 'n_')}"
    )
  ) 

export_to_excel(`numbers` = table_3, `formatted` = table_3_formatted, file = "output/table_3_ring_changes.xlsx")
```


# Table 4: Model parameters

```{r}
# table 4 calculations
alpha_TH_lims <- readr::read_rds("cache/alpha_TH_lims.rds")
calculate_lambda <- function(alpha_TH, k) alpha_TH / (1 - alpha_TH) * k
table_4 <- 
  fits_wide %>%
  mutate(
    alpha_TH_min = alpha_TH_lims[1],
    alpha_TH_max = alpha_TH_lims[2],
    lambda_min = map2_dbl(alpha_TH_min, k, ~calculate_lambda(.x, .y)),
    lambda_max = map2_dbl(alpha_TH_max, k, ~calculate_lambda(.x, .y))
  )

# formatted table
table_4_formatted <- table_4 %>%
  transmute(
    scenario = paste("scenario", row_number()),
    x_fct = as.character(x_fct),
    alpha_W = sprintf("%.1f", alpha_W),
    `alpha_ETC * alpha_NADPH` = sprintf("%.3f \U00B1 %.3f", alpha_NADPH, alpha_NADPH_se),
    `alpha_FAD / alpha_NADPH` = sprintf("%.2f \U00B1 %.2f", alpha_FAD_vs_NADPH, alpha_FAD_vs_NADPH_se),
    alpha_TH = sprintf("%.3f-%.3f", alpha_TH_min, alpha_TH_max),
    lambda = sprintf("%.3f-%.3f", lambda_min, lambda_max)
  ) %>%
  arrange(desc(x_fct), alpha_W) %>%
  pivot_longer(cols = -scenario) %>%
  pivot_wider(values_from = value, names_from = scenario)

# export
export_to_excel(`numbers` = table_4, `formatted` = table_4_formatted, file = "output/table_4_model_fits.xlsx")
```


# Table S1: H accounting

```{r}
# load from model calculations
H_accounting <- readr::read_rds("cache/H_accounting.rds") %>%
  select(-check)

# table S1
table_s1 <- H_accounting %>%
  mutate(
    across(n_A_W:n_x, 
           function(x) x/n_BP, 
           .names = "f_{str_remove(.col, 'n_')}"
    )
  ) 

# export table S1
table_s1_formatted <-
  H_accounting %>%
  transmute(
    scenario = scenario, 
    bp = str_remove(bp_short, "-"),
    across(n_A_W:n_x, 
           function(x) sprintf("%s%%\n(%s)", round(100 * x/n_BP, 2), round(x, 2)), 
           .names = "f_{str_remove(.col, 'n_')}"
    )
  ) 

export_to_excel(`numbers` = table_s1, `formatted` = table_s1_formatted, file = "output/table_S1_H_accounting.xlsx")
```


# Dataset 1: all data

```{r}
samples_w_epsilons %>%
  left_join(compounds, by = "compound") %>%
  left_join(metadata, by = c("organism", "growth")) %>%
  arrange(organism, growth, analysis, compound) %>%
  group_by(organism, growth) %>%
  mutate(
    # renumber analyses for clarity
    analysis = setNames(1:length(unique(analysis)), paste(unique(analysis)))[paste(analysis)],
    n_integrations = as.integer(n_integrations)
  ) %>% 
  ungroup() %>%
  # columns for the export table)
  select(
    growth, TD.hours, 
    analysis, 
    compound = bp_short,
    rt.s = rt, amp2.mV, amp3.mV, 
    area2.Vs, area3.Vs, area_all.Vs,
    n_integrations, rel_abundance,
    raw_d2H.permil, corrected_d2H.permil, corrected_d2H_error.permil,
    d2H_water.permil, eps_2H_lipid_water.permil, eps_2H_lipid_water_error.permil
  ) %>%
  export_to_excel(file = "output/dataset_1_all_data.xlsx")
```
