==> 04_figures.Rmd <==
---
title: "Figures"
author: "Leavitt, Kopf, Weber, Pearson"
date: "Last knitted on `r format(Sys.Date(), '%d %b %Y')`"
output:
  html_document:
    df_print: paged
    number_sections: no
    toc: yes
    toc_float: yes
    toc_depth: 3
    code_folding: show
    standalone: true
editor_options:
  chunk_output_type: console
---

# Setup

```{r "setup", message = FALSE, echo = FALSE}
# packages
library(tidyverse) # general data wrangling and plotting
library(latex2exp) # typesetting of latex math
library(cowplot) # plot arrangement

# scripts
source("scripts/table_functions.R")
source("scripts/plotting_functions.R")

# global knitting options for code rendering
knitr::opts_chunk$set(
  eval = TRUE, echo = FALSE,
  collapse = TRUE, comment = "#>",
  dev = c("png", "pdf", "postscript"),
  dev.args = list(pdf = list(encoding = "WinAnsi", useDingbats = FALSE)),
  fig.keep = "all",
  fig.path = file.path("plots", "")
)
```

> NOTE: please knit the `01_calculations.Rmd` and `02_model.Rmd` notebooks first to prepare the data used for generating the tables.

# Load data

```{r}
# load data
calibration_standards <- readxl::read_excel("data/calibration_standards.xlsx")
internal_standard <- readxl::read_excel("data/internal_standard.xlsx") %>%
  mutate(abs_err = abs(calibrated_d2H.permil - true_d2H.permil)) %>%
  arrange(amp2.mV)
compounds <- readxl::read_excel("data/compounds.xlsx")
metadata <- readxl::read_excel("data/metadata.xlsx")
samples_w_info <- readr::read_rds("cache/samples_w_epsilons.rds") %>%
  left_join(compounds, by = "compound") %>%
  left_join(metadata, by = c("organism", "growth"))
samples_sum_w_info <- readr::read_rds("cache/samples_sum.rds") %>%
  left_join(compounds, by = "compound") %>%
  left_join(metadata, by = c("organism", "growth"))
samples_ab_weighted_sum <- readr::read_rds("cache/samples_ab_weighted_sum.rds")
chromatograms <- readr::read_rds("cache/chromatograms.rds")
growth_regs <- readr::read_rds("cache/growth_regs.rds")
fits_wide <- readr::read_rds("cache/fits_wide.rds")
fits_wide_w_lambdas <- readr::read_rds("cache/fits_wide_w_lambdas.rds")
fits_alphas_sum <- readr::read_rds("cache/fits_alphas_sum.rds")
NADPH_imbalances <- readr::read_rds("cache/NADPH_imbalances.rds")
```

# Plotting constants

```{r}
bp_colors <- c("#D55E00", "#009E73", "#56B4E9", "#E69F00")
bp_shapes <- c(23, 21, 22, 24)
```

# Plots

## Chromatogram

```{r "plot_chrom", fig.width=9, fig.height=5, warning = FALSE}
p_chroms <- chromatograms %>%
  # show only mass 2 trace across the paks rt widow
  filter(mass == "m2", time.s >= 1100, time.s <= 1300) %>%
  ggplot() +
  aes(time.s, signal.mV/1000) + 
  # data
  geom_line(color = "dark gray") +
  geom_ribbon(
    data = function(df) filter(df, !is.na(peak)),
    map = aes(fill = peak, ymin = bgrd2.mV/1000, ymax = signal.mV/1000),
    show.legend = FALSE
  ) +
  # peak labels
  geom_text(
    data = function(df) {
      df %>%
        filter(!is.na(peak)) %>%
        group_by(peak, mass) %>%
        summarize(
          apex.mV = max(signal.mV),
          apex.s = time.s[signal.mV == max(signal.mV)],
          .groups = "drop")
    },
    map = aes(y = apex.mV/1000, x = apex.s, label = peak, color = peak),
    vjust = -0.5, show.legend = FALSE
  ) +
  # ghost points to get legend right
  geom_point(
    data = function(df) df %>% filter(!is.na(peak)) %>% select(peak) %>% unique(),
    map = aes(x = Inf, y = Inf, fill = peak, shape = peak),
    size = 0
  ) +
  # scales
  scale_y_continuous(expand = c(0, 0), breaks = 2*0:5) + expand_limits(y = c(0, 10)) +
  scale_x_continuous(
    breaks = c(19:21) * 60, expand = c(0, 0),
    labels = function(x) sprintf("%.0f s\n%.0f min", x, x/60)
  ) +
  scale_fill_manual(values = c(bp_colors, tail(bp_colors, 1))) +
  scale_color_manual(values = c(bp_colors, tail(bp_colors, 1))) +
  scale_shape_manual(values = c(bp_shapes, tail(bp_shapes, 1))) +
  # themes and labels
  theme_figure(legend = TRUE, grid = FALSE) +
  theme( # to get spacing right for the structures
    legend.key.height = unit(0.16, "npc"),
    legend.spacing = unit(0, "npc"),
    legend.margin = margin(l = 0.1, t = 0, unit = "npc"),
    legend.justification = "top",
    plot.margin = margin(t = 0.1, unit = "npc")
  ) +
  labs(x = "retention time", y = "intensity [V]",
       shape = NULL, color = NULL, fill = NULL) +
  # larger lgend entries
  guides(shape = guide_legend(override.aes = list(size = 4)))

p_chroms
```


## Data overview

```{r "plot_data_overview", fig.width=6, fig.height=6, warning = FALSE}
# plot data frame with shift ranges
plot_df <- 
  samples_sum_w_info %>%
  mutate(
    shift = 3,
    TD.hours_min = min(TD.hours - (1.5 + parse_number(bp_short)) * shift),
    TD.hours_max = max(TD.hours - (1.5 + parse_number(bp_short)) * shift)
  ) %>%
  ungroup() %>%
  mutate(
    eps_value = eps_weighted_mean.permil,
    eps_error = eps_error.permil
  )
plot_reg <- filter(growth_regs, variable == "eps_weighted_mean.permil")

# plot
p_data <- 
  plot_df %>%
  ggplot() +
  # doubling time indicators (vertical dashed lines)
  geom_vline(
    data = function(df) df %>% select(TD.hours) %>% unique(),
    map = aes(xintercept = TD.hours), linetype = 2
  ) + 
  # doubling time group indicators (gray back panels)
  geom_rect(
    data = function(df) df %>% select(TD.hours, shift) %>% unique(),
    map = aes(xmin = TD.hours - 2 * shift, xmax = TD.hours + 1.75 * shift, 
               ymin = -Inf, ymax = Inf),
    fill = "gray", alpha = 0.5
  ) +
  # analyses numbers
  geom_text(
    data = function(df) df %>% select(TD.hours, n_analyses) %>% unique(),
    map = aes(x = TD.hours, y = -Inf, label = sprintf("n = %d", n_analyses)),
    hjust = 0.5, vjust = -0.5, size = 5, fontface = "bold"
  ) +
  # fit line
  geom_line(
    data = function(df) {
      df %>% select(organism, TD.hours_min, TD.hours_max) %>%
        unique() %>%
        left_join(plot_reg, by = "organism") %>%
        mutate(
          fit_line = pmap(list(fit, TD.hours_min, TD.hours_max), function(fit, hrs_min, hrs_max) {
            fit_hrs <- seq(hrs_min, hrs_max, length.out = 10)
            fit_pred <- predict(fit, newdata = data.frame(TD.hours = fit_hrs), se.fit = TRUE)
            tibble(
              TD.hours = fit_hrs,
              eps.permil = fit_pred$fit,
              eps_error.permil = fit_pred$se.fit
            )
          })
        ) %>%
        select(organism, fit_line) %>%
        unnest(fit_line)
    },
    mapping = aes(x = TD.hours, y = eps.permil), size = 1
  ) + 
  # regression line info
  geom_text(
    data = function(df) {
      df %>% select(organism, TD.hours_min, TD.hours_max) %>%
        unique() %>%
        left_join(plot_reg, by = "organism") %>%
        mutate(
          label = sprintf("$\\overset{%.2f \\pm %.2f \U2030\\,/\\,hour}{R^2:\\,%.2f,\\,p-value:\\,%s}$",
                          estimate, std.error, adj.r.squared, format_with_signif(p.value, 1)) %>%
          # label = sprintf("$\\overset{%.2f \\pm %.2f \U2030\\,/\\,hour}{\\[%.2f; %.2f\\]\\,\U2030/hour}$",
          #                 estimate, std.error, lci, uci) %>%
            latex2exp::TeX() %>% as.character()
        )
    },
    mapping = aes(x = 70, y = -270, label = label),
    parse = TRUE, hjust = 0.5, size = 4
  ) + 
  # error bars
  geom_errorbar(
    map = aes(
      x = TD.hours - (1.5 + parse_number(bp_short)) * shift, 
      color = bp_short,
      ymin = eps_value - eps_error, 
      ymax = eps_value + eps_error),
    width = 1, show.legend = FALSE
  ) +
  # scales
  scale_x_continuous(breaks = 0:5*20) + 
  scale_y_continuous(breaks = -15 * 10:40 + 5) +
  scale_fill_manual(values = bp_colors) +
  scale_color_manual(values = bp_colors) +
  scale_shape_manual(values = bp_shapes) +
  expand_limits(y = -295, x = c(20, 100)) +
  # theme
  theme_figure(grid = FALSE) + 
  theme(
    legend.position = "bottom",
    legend.title = element_blank(),
    aspect.ratio = 1
  ) +
  labs(
    x = TeX("$T_D\\,\\[hours\\]$"),
    y = TeX("$^2\\epsilon_{L/W}\\,\\[\U2030\\]$"),
    shape = "biphytane", fill = "biphytane", color = "biphytane"
  ) + 
  # make the legend symbols larger
  guides(
    size = guide_legend(override.aes = list(shape = 21)), 
    shape = guide_legend(override.aes = list(size = 4))
  ) + 
  # individual points
  geom_point(
    map = aes(
      x = TD.hours - (1.5 + parse_number(bp_short)) * shift, 
      y = eps_value, 
      #y = eps_mean.permil,
      shape = bp_short, fill = bp_short),
    size = 5
  ) 

p_data
```

## Model fit parameters

### Alphas

```{r "plot_model_parameters_alphas", fig.width=6.5, fig.height=6, warning=FALSE}
colors <- viridis::plasma(begin = 0.7, end = 0.0, n =3)

p_model_alphas <-
  fits_wide %>%
  ggplot() +
  aes(alpha_NADPH, alpha_FAD_vs_NADPH, color = factor(alpha_W), shape = x_fct, linetype = x_fct) +
  geom_smooth(map = aes(color = NULL), method = "loess", formula = y ~ x, color = "black", span = 2, se = FALSE) +
  geom_errorbar(
    map = aes(ymin = alpha_FAD_vs_NADPH - alpha_FAD_vs_NADPH_se, ymax = alpha_FAD_vs_NADPH + alpha_FAD_vs_NADPH_se),
    width = 0, linetype = 1) +
  geom_errorbarh(
    map = aes(xmin = alpha_NADPH - alpha_NADPH_se, xmax = alpha_NADPH + alpha_NADPH_se),
    height = 0, linetype = 1) +
  geom_point(size = 4) +
  scale_color_manual(values = colors) +
  scale_x_continuous(sec.axis = sec_permil_axis()) +
  scale_y_continuous(sec.axis = sec_permil_axis()) +
  labs(
    x = expression(alpha[ETC] %.% alpha[NADPH]),
    y = expression(alpha[FAD] / alpha[NADPH]),
    color = expression(alpha[W]),#expression(alpha[W] ~"&"~ alpha[ex]),
    shape = expression("H"[2]*"O exch."), 
    linetype = expression("H"[2]*"O exch.")
  ) + 
  theme_figure(grid = FALSE) +
  theme(legend.position = c(0.82, 0.7)) +
  coord_cartesian(ylim = c(0.1, 0.5))
p_model_alphas
```

### Lambda

```{r "plot_model_parameters_lambda_th", fig.width=6, fig.height=6, warning=FALSE}
alpha_TH_lims <- c(0.222, 0.566)
alpha_TH_mid <- 0.5 * sum(alpha_TH_lims)

p_model_lambda_th <- fits_wide_w_lambdas %>%
  filter(alpha_W != 0.7) %>%
  ggplot() +
  aes(alpha_TH, lambda, color = factor(alpha_W), group = paste(alpha_W, x_fct), linetype = x_fct) +
  geom_vline(xintercept = alpha_TH_lims, linetype = 2) +
  geom_vline(xintercept = alpha_TH_mid, linetype = 1, size = 0.2) + 
  geom_ribbon(map = aes(ymin = lambda - lambda_se, ymax = lambda + lambda_se, fill = factor(alpha_W)), alpha = 0.1, color = NA) + 
  geom_line() + 
  geom_rect(data = ~.[1,], ymin = -Inf, ymax = Inf, xmin = -Inf, xmax = alpha_TH_lims[1],
    fill = "white", color = NA, alpha = 0.5) +
  geom_rect(data = ~.[1,], ymin = -Inf, ymax = Inf, xmin = alpha_TH_lims[2], xmax = Inf, 
            fill = "white", color = NA, alpha = 0.5) +
  scale_y_log10() +
  scale_x_continuous(breaks = 0:10 * 0.2, sec.axis = sec_permil_axis(breaks = 0:10 * 0.2)) +
  scale_color_manual(values = colors[c(1,3)]) +
  scale_fill_manual(values = colors[c(1,3)]) +
  coord_cartesian(ylim = c(0.009, 1.1), xlim = c(0.075, 0.98), expand = FALSE) + 
  labs(
    x = expression(alpha[TH]), y = expression(lambda), 
    color = expression(alpha[W]),#expression(alpha[W] ~"&"~ alpha[ex]),
    fill = expression(alpha[W]),#expression(alpha[W] ~"&"~ alpha[ex]),
    linetype = expression("H"[2]*"O exch.")
  ) +
  theme_figure(grid = FALSE) +
  theme(
    legend.position = c(0.85, 0.24),
    legend.spacing = unit(0, "npc"),
    legend.background = element_blank()
  )
p_model_lambda_th
```

## NADH and NADPH

$$
\begin{aligned}
R_{NADH} &= 
  \frac{\alpha_{ETC} \cdot R_W}{\alpha_{TH} + f_L \cdot (1 - \alpha_{TH})} \\
R_{NADPH} &= \alpha_{TH} \cdot R_{NADH}
\end{aligned}
$$

```{r "plot_model_NADH_NADPH", fig.width=8, fig.height=6, warning=FALSE}
# specific example with alpha_TH_mid, alpha_W = 0.9, no water exchange
fits_examples <- 
  fits_wide_w_lambdas %>% 
  filter(alpha_TH == round(alpha_TH_mid, 2), alpha_W == 0.9) %>%
  # assume alpha_NADPH = 0.9
  mutate(alpha_ETC = alpha_NADPH/0.9)

# plot data
fits_examples %>%
  # fractional consumption of NADH
  crossing(fN = seq(0, 1, length.out = 100)) %>%
  mutate(
    fL = 1 - fN,
    alpha_NADH_vs_W = alpha_ETC / (alpha_TH + fL * (1 - alpha_TH)),
    alpha_NADPH_vs_W = alpha_TH * alpha_NADH_vs_W
  ) %>%
  # prep for plotting
  pivot_longer(cols = c(alpha_NADH_vs_W, alpha_NADPH_vs_W)) %>%
  mutate(
    name = name %>% as_factor() %>%
      fct_recode(
        "$R_{NADH}$" = "alpha_NADH_vs_W",
        "$R_{NADPH}$" = "alpha_NADPH_vs_W",
      )
  ) %>%
  # plot
  ggplot() +
  aes(x = fN, y = value, color = name, linetype = x_fct) +
   # range
  annotate(
    "rect",
    xmin = 1 - max(filter(NADPH_imbalances, alpha_TH >= alpha_TH_lims[1], alpha_TH <= alpha_TH_lims[2])$fL), 
    xmax = Inf, ymin = -Inf, ymax = Inf, 
    fill = "gray", alpha = 0.2, color = NA
  ) +
  # data
  geom_line() +
  geom_hline(
    data = . %>% select(x_fct, alpha_ETC) %>% unique() %>% mutate(name = "$\\alpha_{ETC}$"),
    map = aes(yintercept = alpha_ETC, linetype = x_fct, color = name)
  ) + 
  # water
  geom_hline(yintercept = 1, linetype = 1, size = 1) +
  annotate("text", x = 0, y = 1, label = "H[2]*O", hjust = 1.2, vjust = -0.2, parse = TRUE, size = 7) +
  # scales
  #scale_x_reverse(
  scale_x_continuous(
    breaks = 0.2 * (0:10),
    expand = c(0, 0)
    #labels = function(x) sprintf("%s\n(%s %%)", x, -x * 100)
  ) +
  scale_y_continuous(
    breaks = 0.2 * (1:20),
    labels = function(x) sprintf("%+0.f", (x-1) * 1000)
  ) +
  coord_cartesian(ylim = c(0.2, 2)) +
  scale_color_brewer(palette = "Set1", labels = latex_labeller) +
  theme_figure(grid = FALSE) +
  labs(
    #x = "fractional consumption of NADH flux\n(NADPH imbalance flux in %)",
    x = "fractional consumption of NADH flux",
    y = TeX("$\\delta^2H\\,\\[\U2030\\]$"),
    color = "values",
    linetype = expression("H"[2]*"O exch.")
  )

```

## NADPH imbalance

```{r "plot_model_NADPH_imbalance", fig.width=8, fig.height=5, warning=FALSE}
NADPH_imbalance_data <-
  bind_rows(
    NADPH_imbalances %>%
      filter(alpha_TH >= alpha_TH_lims[1], alpha_TH <= alpha_TH_lims[2]) %>%
      group_by(growth) %>%
      summarize(
        balance = mean(NADPH_imb[near(alpha_TH, round(alpha_TH_mid, 2))]),
        balance_min = min(NADPH_imb),
        balance_max = max(NADPH_imb)
      ) %>%
      left_join(samples_ab_weighted_sum, by = "growth") %>%
      mutate(
        eps = eps_ab_weighted_mean.permil,
        eps_se = eps_ab_weighted_mean_error.permil,
        source = "this study"
      ),
    readxl::read_excel("data/literature.xlsx", sheet = "Wijker et al. 2019") %>%
      filter(type == "WT") %>%
      mutate(
        balance = balance_vs_GU,
        balance_min = balance_vs_GU - balance_vs_GU_err,
        balance_max = balance_vs_GU + balance_vs_GU_err,
        # balance = balance_vs_e_flux,
        # balance_min = balance_vs_e_flux - balance_vs_e_flux_err,
        # balance_max = balance_vs_e_flux + balance_vs_e_flux_err,
        source = "Wijker et al. 2019"
      )
  )

NADPH_imbalance_data %>%
  ggplot() +
  aes(balance, eps, ymin = eps - eps_se, ymax = eps + eps_se, 
      xmin = balance_min, xmax = balance_max, color = source) +
  geom_smooth(
    data = . %>% filter(source != "this study"),
    method = "lm", formula = y ~ x, se = FALSE
  ) +
  geom_smooth(
    map = aes(color = NULL), color = "black",
    method = "lm", formula = y ~ x, se = FALSE
  ) +
  geom_errorbar(width = 0) + 
  geom_errorbarh(height = 0) +
  geom_point(size = 4) +
  theme_figure(grid = FALSE) +
  scale_color_manual(values = c("black", "gray60")) +
  scale_x_continuous(labels = function(x) 100 * x) +
  scale_y_continuous(breaks = 100 * c(-5:2)) +
  labs(x = "Predicted NADPH imbalance flux (%)",
       y = TeX("$^2\\epsilon_{L/W}\\,\\[\U2030\\]$")) +
  coord_cartesian(ylim = c(-500, 200))
```


## Data with model fits

```{r "plot_data_with_model", fig.width=6, fig.height=6, warning = FALSE}
# include the model values from the NLS fit
eps_model <-
  fits_alphas_sum %>%
  mutate(
    eps_weighted_mean.permil = (m_alpha - 1) * 1000,
    eps_error.permil = m_alpha_sd * 1000,
  )
eps_model_fits <- 
  eps_model %>%
  filter(TD.hours > min(TD.hours), TD.hours < max(TD.hours))
circle_radius <- 3

# plot
p_data_w_model <- 
  samples_sum_w_info %>%
  ggplot() +
  aes(x = TD.hours, y = eps_weighted_mean.permil, shape = bp_short, color = bp_short, fill = bp_short) +
  # error bars
  geom_errorbar(
    map = aes(ymin = eps_weighted_mean.permil -  eps_error.permil, 
              ymax = eps_weighted_mean.permil +  eps_error.permil),
    width = 1, show.legend = FALSE
  ) +
  # model fits
  geom_line(
    data = eps_model, 
    map = aes(group = paste(growth, bp_short)),
    linetype = 3, size = 1, show.legend = FALSE
  ) +
  # data points
  geom_point(size = 5, color = "black") +
  # model fits circles
  geom_path(
    data = eps_model_fits %>%
      mutate(id = row_number()) %>%
      crossing(t = seq(0, 2 * pi, length.out = 100)) %>%
      mutate(
          eps_weighted_mean.permil = eps_weighted_mean.permil +
            circle_radius * 63/80 * sin(t),
          TD.hours = TD.hours + circle_radius * cos(t)
      ),
    map = aes(group = id), size = 1, show.legend = FALSE
  ) +
  # scales
  scale_fill_manual(values = bp_colors) +
  scale_color_manual(values = bp_colors) +
  scale_shape_manual(values = bp_shapes) +
  scale_y_continuous(breaks = -15 * 10:40 + 5) +
  scale_x_continuous(breaks = 0:6*20, expand = c(0, 0)) +
  expand_limits(y = -295) +
  coord_cartesian(xlim = c(19, 102)) +
  # theme
  theme_figure(grid = FALSE) + 
  theme(
    legend.position = "bottom",
    legend.title = element_blank(),
    aspect.ratio = 1
  ) +
  labs(
    x = TeX("$T_D\\,\\[hr\\]$"),
    y = TeX("$^2\\epsilon_{L/W}\\,\\[\U2030\\]$"),
    shape = "biphytane", fill = "biphytane", color = "biphytane"
  ) +
  # make the legend symbols larger
  guides(
    size = guide_legend(override.aes = list(shape = 21)), 
    shape = guide_legend(override.aes = list(size = 4))
  ) 

p_data_w_model
```

## Peak size effect

```{r "peak_size_effect", fig.width = 10, fig.height = 6, warning=FALSE}
derived_params <- read_rds("cache/derived_params.rds")

p_peak_size <- derived_params %>%
  filter(mean_amp < 20) %>%
  mutate(y = `bgrd / peak`, y_err = `bgrd / peak error`) %>%
  select(analysis, y, y_err, ends_with("amp"), ends_with("area")) %>% 
  pivot_longer(
    cols = c(ends_with("area"), ends_with("amp")), 
    names_to = c("name", "variable"), names_pattern = "(.*)_(.*)"
  ) %>% 
  pivot_wider() %>%
  mutate(
    variable = as_factor(variable) %>% 
      fct_recode("mean amplitude (A) in V" = "amp", ",mean area (A) in Vs" = "area")
  ) %>%
  ggplot() +
  aes(x = mean, y = y) +
  geom_errorbar(map = aes(ymin = y - y_err, ymax = y + y_err), width = 0) + 
  geom_errorbarh(map = aes(xmin = min, xmax = max), height = 0) +
  geom_point(size = 4) +
  facet_grid(~variable, scales = "free_x", switch = "x") + 
  scale_y_continuous(labels = function(x) sprintf("%.0f%%", 100 * x)) +
  theme_figure() +
  theme(
    strip.placement = "outside",
    strip.background = element_blank()
  ) +
  labs(x = NULL, y = latex2exp::TeX("$\\frac{A_{bgrd}}{A_{peak}}$"))

p_peak_size
```

## Data calibration summary

```{r "plot_calib_summary", fig.width = 8, fig.height = 7, warning=FALSE}
# load data unique to this plot
calibration_standards <- readxl::read_excel("data/calibration_standards.xlsx")

# root means square error
rmse <- calibration_standards %>%
  group_by(calibration) %>%
  summarize(
    n = n(), 
    rmse = sqrt(mean((calibrated_d2H.permil - true_d2H.permil)^2, na.rm = TRUE)) %>%
      round(2),
    .groups = "drop"
  )
rmse %>% knitr::kable()

# plot
p_calibs <- 
  calibration_standards %>%
  mutate(
    panel = as_factor(calibration) %>%
      fct_relevel("model.simple") %>%
      fct_recode(
        "$\\delta^2H_{raw} = \\beta_0 + \\beta_1 \\delta^2H_{calib}$" = "model.simple",
        "$\\delta^2H_{raw} = \\beta_0 + \\beta_1 \\delta^2H_{calib} + \\beta_2 A$" = "model.amp",
        "$\\delta^2H_{raw} = \\beta_0 + \\beta_1 \\delta^2H_{calib} + \\beta_2 A + \\beta_3 \\sqrt{A}$" = "model.amp.amp.sqrt",
        "$\\delta^2H_{raw} = \\beta_0 + \\beta_1 \\delta^2H_{calib} + \\beta_2 A + \\beta_3 \\delta^2H_{calib} \\sqrt{A}$" = "model.amp.amp.sqrt.cross"
      )
  ) %>%
  ggplot() + 
  aes(amp2.mV/1000, residual_d2H.permil, color = true_d2H.permil) +
  geom_point(alpha = 0.5) +
  facet_wrap(~panel, labeller = latex_labeller) +
  scale_x_log10() + 
  scale_color_continuous(labels = function(x) paste0(x, "\U2030")) +
  coord_cartesian(xlim = c(0.4, 50)) +
  theme_figure() +
  theme(
    aspect.ratio = 1,
    strip.text = element_text(size = 10)
  )  +
  geom_hline(yintercept = 0, linetype = "dotted", color ="gray30") +
  geom_text(
    data = function(df) {
      df %>% select(calibration, panel) %>%
        unique() %>%
        left_join(rmse, by = "calibration")
    },
    mapping = aes(x = Inf, y = Inf, color = NULL, label = sprintf("RMSE: %.1f\U2030", rmse)),
    hjust = 1.1, vjust = 1.5
  ) +
  labs(
    color = TeX("$\\delta^2H$"),
    x = "mass 2 amplitude (A) in V",
    y = TeX("standard ($nC_{15-30}$) residuals in \U2030")
  )

p_calibs
```

## Absolute error estimates

```{r "plot_abs_errors", fig.width = 8, fig.height = 6, warning=FALSE}
# create estimator for absolute error based on the internal standard
abs_err_model <- loess(
  formula = abs_err ~ amp2.mV,
  data = internal_standard,
  span = 0.75
)
calculate_abs_error <- function(amp2.mV, m = abs_err_model) {
  predict(m, newdata = tibble(amp2.mV = amp2.mV))
}

# plot
p_errors <- 
  internal_standard %>%
  ggplot() +
  aes(amp2.mV/1000, abs_err) + 
  # fit line
  geom_line(
    data = function(df) {
      tibble(
        amp2.mV = seq(min(df$amp2.mV), max(df$amp2.mV), length.out = 100),
        abs_err = calculate_abs_error(amp2.mV)
      ) %>% filter(!is.na(abs_err))
    },
    color = "red", size = 1
  ) +
  geom_point(size = 3) +
  theme_figure(grid = TRUE) + 
  # scales
  scale_y_continuous(breaks = c(0, 5, 10, 20, 50, 100)) + 
  scale_x_continuous(breaks = c(0, 1, 2, 5, 10, 15, 20), expand = c(0, 0.5)) +
  #scale_x_log10() + # log scale instead
  labs(
    x = "mass 2 amplitude [V]",
    y = TeX("internal standard ($nC_{36}$) absolute error [\U2030]")
  )

p_errors
```

## All isotope data

### A: data plot

```{r "plot_all_data_points", fig.width=8, fig.height=8, warning=FALSE}
# data plot
p_all_data_points <- 
  samples_w_info %>%
  ggplot() +
  aes(amp2.mV/1000, corrected_d2H.permil, fill = bp_short, shape = bp_short) + 
  # errrors
  geom_errorbar(
    width = 0,
    map = aes(ymin = corrected_d2H.permil - corrected_d2H_error.permil, 
              ymax = corrected_d2H.permil + corrected_d2H_error.permil)
  ) +
  # data
  geom_point(size = 3) +
  # scales
  scale_x_log10() +
  scale_fill_manual(values = bp_colors) +
  scale_color_manual(values = bp_colors) +
  scale_shape_manual(values = bp_shapes) +
  # wrap
  facet_grid(growth~., labeller = latex_labeller, scale = "free_y") + 
  # themes
  theme_figure() +
  theme(legend.position = "bottom", legend.direction = "horizontal") +
  labs(
    x = "mass 2 amplitude [V]",
    y = TeX("corrected $\\delta^2H\\,\\[\U2030\\]$"),
    shape = NULL, fill = NULL
  ) 

p_all_data_points
```

### B: density plot

```{r "plot_all_data_density", fig.width=6, fig.height=4, warning = FALSE}
# density plot
p_all_data_density <- 
  samples_w_info %>%
  ggplot() +
  aes(y = corrected_d2H.permil, color = compound) +
  # data
  geom_density(
    map = aes(x = after_stat(ndensity)),
    alpha = 0.2,
    position = "stack",
    show.legend = FALSE
  ) +
  # include error bars to scale the same way as the main data
  geom_errorbar(
    map = aes(
      ymin = corrected_d2H.permil - corrected_d2H_error.permil,
      ymax = corrected_d2H.permil + corrected_d2H_error.permil,
      x = 0),
    width = 0, size = 0
  ) +
  # base line
  geom_vline(xintercept = 0) +
  # weighted averages (1/err^2 weighted)
  geom_hline(
    data = function(df) {
      df %>% 
        group_by(organism, growth, compound) %>%
        summarize(
          mean = Hmisc::wtd.mean(corrected_d2H.permil, 1/corrected_d2H_error.permil^2),
          .groups = "drop"
        )
    },
    map = aes(yintercept = mean, color = compound), size = 2
  ) +
  # facet
  facet_grid(growth~compound, scales = "free_y") +
  # scales
  scale_x_continuous(exp = c(0, 0)) +
  scale_color_manual(values = bp_colors) +
  # theme
  theme_figure(grid = FALSE) +
  theme(
    axis.line = element_blank(),
    axis.ticks = element_blank(),
    axis.text = element_blank(),
    axis.title = element_blank(),
    strip.background = element_blank(),
    strip.text = element_text(color = "white"),
    panel.border = element_blank(),
    plot.margin = margin(l = -3)
  )+
  labs(
    y = TeX("corrected $\\delta^2H\\,\\[\U2030\\]$"),
    x = "probability density"
  )

p_all_data_density
```

### A & B combined

```{r "plot_all_data_combined", fig.width=10, fig.height=6, warning = FALSE}
# assemble plot
p_all_data <-
  plot_grid(
    p_all_data_points + theme(plot.margin = margin(r = 0)), 
    p_all_data_density + theme(legend.position = "none"),
    align = "h", nrow = 1,
    axis = "tb", rel_widths = c(2, 1)
  )
p_all_data
```


## Relative abundances

```{r "plot_rel_abundances", fig.width=6, fig.height=5}
p_relab <- 
  samples_sum_w_info %>%
  ggplot() +
  aes(growth, rel_abundance_mean, fill = bp_short) +
  geom_bar(stat = "identity", position = "stack", width = 0.9) +
  # scales
  scale_y_continuous(expand = c(0, 0), labels = function(x) paste0(100*x, "%")) +
  scale_x_discrete(expand = c(0, 0)) +
  scale_fill_manual(values = bp_colors) +
  # theme
  theme_figure(grid = FALSE) +
  theme(plot.margin = margin(t = 0.1, unit = "npc")) + 
  labs(
    x = "growth rate",
    y = "relative abundance",
    shape = NULL, fill = NULL, color = NULL
  )
p_relab
```

# Figures

## Figure 1

```{r "figure_1", fig.width=6, fig.height=9, warning = FALSE}
# combine data overview with structures
plot_grid(
  p_data,
  ggdraw() + draw_image("figures/biphytanes.pdf"),
  ncol = 1, rel_heights = c(2, 0.8),
  labels = c("A", "B")
)
```

## Figure 4

```{r "figure_4", fig.width=12.5, fig.height=6, warning = FALSE}
# combine data overview with structures
plot_grid(
  p_model_alphas, p_model_lambda_th,
  nrow = 1, rel_widths = c(1, 0.9),
  labels = c("A", "B")
)
```

## Figure 5

```{r "figure_5", fig.width=6, fig.height=6, warning = FALSE}
p_data_w_model
```

## Figure S1

```{r "figure_S1", fig.width=10, fig.height=12, warning=FALSE}
# combine all data with chroms and rel. abundances
plot_grid(
  p_all_data,
  plot_grid(
    p_chroms + theme(legend.position = "none"),
    p_relab + theme(legend.position = "none", plot.margin = margin(0, 0, 0, 0.05, "npc")),
    nrow = 1, rel_widths = c(1, 1), align = "h", axis = "tb",
    labels = c("B", "C")
  ),
  ncol = 1, rel_heights = c(3, 2),
  labels = c("A", "")
)

```

## Figure S2

```{r "figure_S2", fig.width = 8, fig.height = 10.5, warning=FALSE}
# combine all data with chroms and rel. abundances
plot_grid(
  p_peak_size,
  p_calibs + theme(legend.position = "left"),
  ncol = 1, rel_heights = c(4, 9),
  labels = c("A", "B")
)
```

## Figure S3

```{r "figure_S3", fig.width = 8, fig.height = 6, warning=FALSE}
p_errors
```
